Refactor User classes
