+++ WebApp/src/uk/ac/exeter/QuinCe/User/User.java
@@ -1,4 +1,4 @@
-package uk.ac.exeter.QuinCe.data;
+package uk.ac.exeter.QuinCe.User;
 
 import java.sql.Timestamp;
