+++ WebApp/src/uk/ac/exeter/QuinCe/jobs/user/SendEmailVerificationMailJob.java
@@ -5,8 +5,8 @@ import java.util.Properties;
 
 import org.apache.commons.mail.EmailException;
 
-import uk.ac.exeter.QuinCe.data.User;
-import uk.ac.exeter.QuinCe.database.User.UserDB;
+import uk.ac.exeter.QuinCe.User.User;
+import uk.ac.exeter.QuinCe.User.UserDB;
 import uk.ac.exeter.QuinCe.jobs.InvalidJobParametersException;
 import uk.ac.exeter.QuinCe.jobs.Job;
 import uk.ac.exeter.QuinCe.jobs.JobFailedException;
