+++ WebApp/src/uk/ac/exeter/QuinCe/web/User/AuthenticatedFilter.java
@@ -16,7 +16,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
-import uk.ac.exeter.QuinCe.data.User;
+import uk.ac.exeter.QuinCe.User.User;
 
 /**
  * Filter for all URLs, to determine whether or not the user is logged in.
