+++ WebApp/src/uk/ac/exeter/QuinCe/web/User/JobsPermissionFilter.java
@@ -2,7 +2,7 @@ package uk.ac.exeter.QuinCe.web.User;
 
 import javax.servlet.annotation.WebFilter;
 
-import uk.ac.exeter.QuinCe.data.User;
+import uk.ac.exeter.QuinCe.User.User;
 
 @WebFilter("/jobs/*")
 public class JobsPermissionFilter extends PermissionsFilter {
