+++ WebApp/src/uk/ac/exeter/QuinCe/web/User/SendEmailCodeBean.java
@@ -3,8 +3,8 @@ package uk.ac.exeter.QuinCe.web.User;
 import java.util.HashMap;
 import java.util.Map;
 
-import uk.ac.exeter.QuinCe.data.User;
-import uk.ac.exeter.QuinCe.database.User.UserDB;
+import uk.ac.exeter.QuinCe.User.User;
+import uk.ac.exeter.QuinCe.User.UserDB;
 import uk.ac.exeter.QuinCe.jobs.JobManager;
 import uk.ac.exeter.QuinCe.jobs.user.SendEmailVerificationMailJob;
 import uk.ac.exeter.QuinCe.web.BaseManagedBean;
