+++ WebApp/src/uk/ac/exeter/QuinCe/web/User/VerifyEmailBean.java
@@ -1,6 +1,6 @@
 package uk.ac.exeter.QuinCe.web.User;
 
-import uk.ac.exeter.QuinCe.database.User.UserDB;
+import uk.ac.exeter.QuinCe.User.UserDB;
 import uk.ac.exeter.QuinCe.utils.MissingParamException;
 import uk.ac.exeter.QuinCe.web.BaseManagedBean;
 import uk.ac.exeter.QuinCe.web.system.ServletUtils;
